+++ test-data/test-files/paredit_sandbox.clj
@@ -109,6 +109,25 @@ string. "
 ; Should not delete `#`
 (#|())
 
+;Should delete entire #() form with backspace and delete
+#()
+
+; Should not delete #  
+#{|21 2}
+
+; Should delete #
+#|{21}
+
+;; Backspace should delete the #
+[|#(prn "hello")]
+|#(prn "hello"|)
+
+;; Delete should delete space and #
+[# |(prn "hello")]
+
+; Delete should move before # without deleting it
+(#(|inc 2))
+
 ;; Pair selecting
 
 (cond
@@ -125,15 +144,21 @@ string. "
   false (assoc :b 2)
   true  (assoc :c 3))
 
+(case 1
+  1 "one"
+  2 "two"
+  3 "three"
+  "other")
 
+(condp = 2
+  1 "one"
+  2 "two"
+  3 "three"
+  "other")
 
-;; cond pair
-
-(cond
-  (= 1 2) 1
-  (= 2 2) 2
-  :else 23)
-
+(condp some [1 2 3 4]
+  #{0 6 7} :>> inc
+  #{5 9}   :>> dec)
 
 ;; === Slurp Backward
 
@@ -173,3 +198,61 @@ foo [|]
             e
             f) g]
   :a)
+
+;; === Forward Slurp
+;; Issue: Strings are treated as lists, causing unwanted space insertion
+
+;; Slurp Sexp Forward with string at cursor start
+;; Before: "|"somestuff
+;; Expected: "|somestuff"
+""somestuff
+
+;; Slurp Sexp Forward with string and following form
+;; Before: "|"  (some-form other-args)
+;; Expected: "|(some-form other-args)"
+"|"  (some-form other-args)
+
+
+;; Empty list slurp forward - should NOT add space
+;; Before: (|) "foo"
+;; Expected: (|"foo")
+(|) "foo"
+
+;; Whitespace-only list slurp forward
+;; Before: (|   ) "foo"
+;; Expected: (|"foo")
+(|   ) "foo"
+
+;; Non-empty list slurp forward - should add space
+;; Before: "aaa|" ccc
+;; Expected: "aaa| ccc"
+"aaa|" ccc
+
+;; === Slurp Forward with Data Structures
+
+;; Empty vector slurp forward
+;; Before: [|] 1
+;; Expected: [|1]
+[|] 1
+
+;; Empty map slurp forward
+;; Before: {|} :key
+;; Expected: {|:key}
+{} :key
+
+;; Empty set slurp forward
+;; Before: #{|} value
+;; Expected: #{|value}
+#{|} value
+
+;; Non-empty vector with space
+;; Before: [a|] b
+;; Expected: [a| b]
+[a] b
+
+;; Nested structure slurp forward
+;; Command: Paredit Slurp Forward (twice) 
+;; Before: ([|]) "nested"
+;; Expected: ([|"nested"])
+([|]) "nested" ;;->> ([|] "nested")->>([|] "nested")->>[(|"nested")] 
+_